requirements.txt
streamlit==1.28.0
pandas==2.1.0
numpy==1.24.0
pydantic==2.4.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
mistralai==0.0.11
anthropic==0.7.0
openai==1.3.0
plotly==5.17.0
streamlit-extras==0.3.0
pdfplumber==0.9.0
python-docx==0.8.11
